requirements.txt
pandas>=1.5.0
numpy>=1.21.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pydantic>=1.8.0
python-multipart>=0.0.5

# Streamlit Web Interface
streamlit>=1.28.0
plotly>=5.15.0